refactor step-ids to file
