+++ engine/steps/aave/deploy/01-deploy-supply.ts
@@ -1,5 +1,5 @@
 import { DeployFunction } from 'hardhat-deploy/types'
-import { STEP_TYPE_ID_AAVE_SUPPLY } from '../tslib/supply-helper'
+import { STEP_TYPE_ID_AAVE_SUPPLY } from '../../step-ids'
 import { getWrappedNativeAddress } from '@freemarket/step-sdk'
 import { deployStep } from '@freemarket/step-sdk/tslib/deploy-step'
 import { getPoolAddress } from '../tslib/getPoolAddress'
