+++ engine/steps/aave/test/borrow-test.ts
@@ -5,7 +5,8 @@ import { AaveSupplyHelper } from '../tslib/supply-helper'
 import { ADDRESS_ZERO, EncodingContext } from '@freemarket/core'
 import { getTestFixture, MockWorkflowInstance, validateAction, getUsdc, printEvents } from '@freemarket/step-sdk/tslib/testing'
 import { AaveBorrow, AaveSupply } from '../tslib/model'
-import { AaveBorrowHelper, STEP_TYPE_ID_AAVE_BORROW } from '../tslib/borrow-helper'
+import { AaveBorrowHelper } from '../tslib/borrow-helper'
+import { STEP_TYPE_ID_AAVE_BORROW } from '../../step-ids'
 import { formatNumber } from '@freemarket/step-sdk'
 import { Eip1193Bridge } from '@ethersproject/experimental'
 import { Signer } from '@ethersproject/abstract-signer'
