+++ engine/steps/aave/test/supply-test.ts
@@ -1,7 +1,8 @@
 import { expect } from 'chai'
 import hardhat, { ethers, deployments } from 'hardhat'
 import { AaveSupplyAction, IAaveV3Pool__factory, IERC20__factory } from '../typechain-types'
-import { AaveSupplyHelper, STEP_TYPE_ID_AAVE_SUPPLY } from '../tslib/supply-helper'
+import { AaveSupplyHelper } from '../tslib/supply-helper'
+import { STEP_TYPE_ID_AAVE_SUPPLY } from '../../step-ids'
 import { ADDRESS_ZERO, EncodingContext } from '@freemarket/core'
 import { getTestFixture, MockWorkflowInstance, validateAction, getUsdt, confirmTx } from '@freemarket/step-sdk/tslib/testing'
 import { AaveSupply } from '../tslib/model'
