+++ engine/steps/base-branches/tslib/AssetBalanceBranchHelper.ts
@@ -2,8 +2,7 @@ import type { AssetBalanceBranch } from './model'
 import { EncodedWorkflowStep, EncodingContext } from '@freemarket/core'
 
 import { AssetComparisonHelperBase } from './AssetComparisonHelperBase'
-
-export const STEP_TYPE_ID_ASSET_BALANCE_BRANCH = 2
+import { STEP_TYPE_ID_ASSET_BALANCE_BRANCH } from '../../step-ids'
 
 export class AssetBalanceBranchHelper extends AssetComparisonHelperBase {
   async encodeWorkflowStep(context: EncodingContext<AssetBalanceBranch>): Promise<EncodedWorkflowStep> {
