+++ engine/steps/base-branches/tslib/PreviousOutputBranchHelper.ts
@@ -2,8 +2,7 @@ import type { PreviousOutputBranch } from './model'
 import { EncodedWorkflowStep, EncodingContext } from '@freemarket/core'
 
 import { AssetComparisonHelperBase } from './AssetComparisonHelperBase'
-
-export const STEP_TYPE_ID_PREV_OUTPUT_BRANCH = 3
+import { STEP_TYPE_ID_PREV_OUTPUT_BRANCH } from '../../step-ids'
 
 export class PreviousOutputBranchHelper extends AssetComparisonHelperBase {
   async encodeWorkflowStep(context: EncodingContext<PreviousOutputBranch>): Promise<EncodedWorkflowStep> {
