+++ engine/steps/step-ids.ts
@@ -1,0 +1,19 @@
+export const STEP_TYPE_ID_AAVE_REPAY = 111;
+export const STEP_TYPE_ID_AAVE_SUPPLY = 102;
+export const STEP_TYPE_ID_AAVE_BORROW = 110;
+export const STEP_TYPE_ID_ADD_ASSET = 100;
+export const STEP_TYPE_ID_CHAIN_BRANCH = 1;
+export const STEP_TYPE_ID_ASSET_BALANCE_BRANCH = 2;
+export const STEP_TYPE_ID_PREV_OUTPUT_BRANCH = 3;
+export const STEP_TYPE_ID_CURVE = 103;
+export const STEP_TYPE_ID_LIDO_ETH_TO_STETH = 205
+export const STEP_TYPE_ID_LIDO_STETH_TO_WSTETH = 206
+export const STEP_TYPE_ID_LIDO_WSTETH_TO_STETH = 207;
+export const STEP_TYPE_ID_PENDLE_TOKEN_TO_PT = 301;
+export const STEP_TYPE_ID_RENZO_ETH_TO_EZETH = 204;
+export const STEP_TYPE_ID_STARGATE_BRIDGE = 101;
+export const STEP_TYPE_ID_UNISWAP_EXACT_IN = 104;
+export const STEP_TYPE_ID_UNISWAP_EXACT_OUT = 109;
+export const STEP_TYPE_ID_UNWRAP_NATIVE = 106;
+export const STEP_TYPE_ID_WRAP_NATIVE = 105
+
