+++ engine/steps/wrapped-native/deploy/01-deploy-unwrap-native.ts
@@ -1,7 +1,7 @@
 import { DeployFunction } from 'hardhat-deploy/types'
 import { deployStep } from '@freemarket/step-sdk/tslib/deploy-step'
 import { getWrappedNativeAddress } from '@freemarket/step-sdk/tslib/getWrappedNativeAddress'
-import { STEP_TYPE_ID_UNWRAP_NATIVE } from '../tslib/unwrap-native-helper'
+import { STEP_TYPE_ID_UNWRAP_NATIVE } from '../../step-ids'
 
 const func: DeployFunction = async function (hardhatRuntimeEnv) {
   const chainId = await hardhatRuntimeEnv.getChainId()
