+++ engine/steps/wrapped-native/test/test.ts
@@ -1,13 +1,14 @@
 import { expect } from 'chai'
 import hardhat, { ethers, deployments } from 'hardhat'
-import { STEP_TYPE_ID_WRAP_NATIVE, WrapNativeHelper } from '../tslib/wrap-native-helper'
+import { WrapNativeHelper } from '../tslib/wrap-native-helper'
+import { STEP_TYPE_ID_UNWRAP_NATIVE, STEP_TYPE_ID_WRAP_NATIVE } from '../../step-ids'
 import { ADDRESS_ZERO, createStandardProvider, EncodingContext, IERC20__factory, WORKFLOW_END_STEP_ID } from '@freemarket/core'
 import { confirmTx, getTestFixture, MockWorkflowInstance, validateAction } from '@freemarket/step-sdk/tslib/testing'
 import { TestErc20__factory } from '@freemarket/step-sdk'
 import { UnwrapNativeAction, Weth__factory, WrapNativeAction } from '../typechain-types'
 import { getWrappedNativeAddress } from '@freemarket/step-sdk/tslib/getWrappedNativeAddress'
 import { IERC20 } from '@freemarket/runner'
-import { STEP_TYPE_ID_UNWRAP_NATIVE, UnwrapNative, UnwrapNativeHelper, WrapNative } from '../tslib'
+import { UnwrapNative, UnwrapNativeHelper, WrapNative } from '../tslib'
 import { ContractTransaction } from 'ethers'
 import { TransactionReceipt } from '@ethersproject/providers'
 import { WorkflowStruct } from '@freemarket/core/typechain-types/contracts/IWorkflowRunner'
